+++ client/app/pods/components/select-2-multiple/component.js
@@ -7,3 +7,4 @@ export default Select2Component.extend({
     this.$().select2('val', (this.get('selectedData') || []).mapBy('id'));
   })
 });
+