+++ client/tests/factories/register-decision-task.js
@@ -1,0 +1,9 @@
+import FactoryGuy from 'ember-data-factory-guy';
+
+FactoryGuy.define('register-decision-task', {
+  default: {
+    title: 'Register Decision',
+    type: 'RegisterDecisionTask',
+    completed: false
+  }
+});