+++ engines/tahi_standard_tasks/app/controllers/tahi_standard_tasks/register_decision_controller.rb
@@ -4,20 +4,19 @@ module TahiStandardTasks
       task = Task.find(params[:id])
       requires_user_can :register_decision, task.paper
 
-      if !task || !task.paper.submitted?
-        render json: {
-          errors: { 'errors': ['Invalid Task and/or Paper'] }
-        }, status: 422
-
-      elsif not task.latest_decision_ready?
+      if !task.paper.submitted?
+        task.errors.add(:errors, "Paper is not submitted")
+        fail ActiveRecord::RecordInvalid, task
+      elsif !task.latest_decision_ready?
         render json: {
           errors: { 'errors': ['You must register a verdict first.'] }
         }, status: 422
-
       else
         decision = task.latest_decision
         task.complete_decision
-        task.send_email
+        to_field = task.answer_for('register_decision_questions--to-field').try(:value)
+        subject = task.answer_for('register_decision_questions--subject-field').try(:value)
+        task.send_email(to_field: to_field, subject_field: subject)
         Activity.decision_made! decision, user: current_user
         render json: decision, status: :created
       end