+++ engines/tahi_standard_tasks/app/mailers/tahi_standard_tasks/register_decision_mailer.rb
@@ -6,13 +6,14 @@ module TahiStandardTasks
 
     default from: ENV.fetch('FROM_EMAIL', 'no-reply@example.com')
 
-    def notify_author_email(decision_id:)
+    def notify_author_email(to_field:, subject_field:, decision_id:)
       @decision = Decision.find(decision_id)
       @paper = @decision.paper
-      @recipient = User.find(@paper.creator.id)
+      @recipient_email = to_field || @paper.creator.email
+      @subject = subject_field || "A decision has been registered on the manuscript, \"#{@paper.display_title}\""
 
-      mail(to: @recipient.email,
-           subject: "A decision has been registered on the manuscript, \"#{@paper.display_title}\"")
+      mail(to: @recipient_email,
+           subject: @subject)
     end
   end
 end