+++ engines/tahi_standard_tasks/app/models/tahi_standard_tasks/register_decision_task.rb
@@ -6,6 +6,9 @@ module TahiStandardTasks
 
     # TODO: move these attributes from paper to this task model (https://www.pivotaltracker.com/story/show/84690814)
     delegate :decision_letter, :decision_letter=, to: :paper, prefix: :paper
+
+    delegate :letter_templates, to: :journal
+
     before_save { paper.save! }
 
     def self.permitted_attributes
@@ -27,138 +30,28 @@ module TahiStandardTasks
       DecisionReviser.new(self, decision).process! if decision.revision?
     end
 
-    def send_email
+    def send_email(to_field:, subject_field:)
       RegisterDecisionMailer.delay.notify_author_email(
+        to_field: EmailService.new(email: to_field).valid_email_or_nil,
+        subject_field: subject_field,
         decision_id: paper.decisions.completed.latest)
     end
 
     def send_emails
     end
 
-    # These methods are a bunch of english text. They should be moved to
-    # their own file, but we're not sure where. They're here, instead of a
-    # mailer template, because users can edit the text before it gets
-    # sent out.
-    # rubocop:disable Metrics/LineLength
-    def accept_letter
-      template = <<-TEXT.strip_heredoc
-        Dear Dr. %{author_last_name},
-
-        I am pleased to inform you that your manuscript, %{manuscript_title}, has been deemed suitable for publication in %{journal_name}. Congratulations!
-
-        Your manuscript will now be passed on to our Production staff, who will check your files for correct formatting and completeness. During this process, you may be contacted to make necessary alterations to your manuscript, though not all manuscripts require this.
-
-        If you or your institution will be preparing press materials for this manuscript, you must inform our press team in advance. Your manuscript will remain under a strict press embargo until the publication date and time.
-
-        Please contact me if you have any other questions or concerns. Thank you for submitting your work to %{journal_name}.
-
-        With kind regards,
-
-        [YOUR NAME]
-        %{journal_name}
-      TEXT
-
-      template % template_data
-    end
-
-    def minor_revision_letter
-      template = <<-TEXT.strip_heredoc
-        THIS IS THE MINOR REVISION
-        Dear Dr. %{author_last_name},
-
-        Thank you for submitting your manuscript, %{manuscript_title} to %{journal_name}. After careful consideration, we feel that it has merit, but is not suitable for publication as it currently stands. Therefore, my decision is "Minor Revision."
-
-        We invite you to submit a revised version of the manuscript that addresses the points below:
-
-        ***
-
-        ACADEMIC EDITOR:
-
-        PLEASE INSERT COMMENTS HERE GIVING CONTEXT TO THE REVIEWS AND EXPLAINING WHICH REVIEWER COMMENTS MUST BE ADDRESSED.
-
-        ***
-
-        We encourage you to submit your revision within forty-five days of the date of this decision.
-
-        When your files are ready, please submit your revision by logging on to #{url_for(:root)} and following the instructions for resubmission. Do not submit a revised manuscript as a new submission.
-
-        If you choose not to submit a revision, please notify us.
-
-        Yours sincerely,
-
-        [YOUR NAME]
-        %{journal_name}
-      TEXT
-
-      template % template_data
-    end
-
-    def major_revision_letter
-      template = <<-TEXT.strip_heredoc
-        Dear Dr. %{author_last_name},
-
-        Thank you for submitting your manuscript, %{manuscript_title} to %{journal_name}. After careful consideration, we feel that it has merit, but is not suitable for publication as it currently stands. Therefore, my decision is "Major Revision."
-
-        We invite you to submit a revised version of the manuscript that addresses the points below:
-
-        ***
-
-        ACADEMIC EDITOR:
-
-        PLEASE INSERT COMMENTS HERE GIVING CONTEXT TO THE REVIEWS AND EXPLAINING WHICH REVIEWER COMMENTS MUST BE ADDRESSED.
-
-        ***
-
-        We encourage you to submit your revision within forty-five days of the date of this decision.
-
-        When your files are ready, please submit your revision by logging on to #{url_for(:root)} and following the instructions for resubmission. Do not submit a revised manuscript as a new submission.
-
-        If you choose not to submit a revision, please notify us.
-
-        Yours sincerely,
-
-        [YOUR NAME]
-        %{journal_name}
-      TEXT
-
-      template % template_data
-    end
-
-    def reject_letter
-      template = <<-TEXT.strip_heredoc
-        Dear Dr. %{author_last_name},
-
-        Thank you for submitting your manuscript, %{manuscript_title}, to %{journal_name}. After careful consideration, we have decided that your manuscript does not meet our criteria for publication and must therefore be rejected.
-
-        Specifically:
-
-        ***
-
-        ACADEMIC EDITOR:
-
-        PLEASE INSERT COMMENTS HERE GIVING CONTEXT TO THE REVIEWS AND EXPLAINING HOW THE MANUSCRIPT DOES NOT MEET OUR PUBLICATION CRITERIA.
-
-        ***
-
-        I am sorry that we cannot be more positive on this occasion, but hope that you appreciate the reasons for this decision.
-
-        Yours sincerely,
-
-        [YOUR NAME]
-        %{journal_name}
-      TEXT
-
-      template % template_data
-    end
-    # rubocop:enable Metrics/LineLength
-
     private
+
+    def journal_name
+      @journal_name ||= paper.journal.name
+    end
 
     def template_data
       {
         author_last_name: paper.creator.last_name,
         manuscript_title: paper.display_title(sanitized: false),
-        journal_name: paper.journal.name
+        journal_name: paper.journal.name,
+        your_name: '[YOUR NAME]'
       }
     end
   end