+++ engines/tahi_standard_tasks/spec/mailers/tahi_standard_tasks/register_decision_mailer_spec.rb
@@ -20,26 +20,41 @@ describe TahiStandardTasks::RegisterDecisionMailer do
 
   let(:decision) {
     paper.decisions.create!(
-      letter: "Body text of a Decision Letter",
+      letter: "Body text of a Decision Letter << We Accept",
       verdict: "accept"
     )
   }
 
-  let(:email) {
-    described_class.notify_author_email(decision_id: decision.id)
+  let(:email_with_no_fields_specified) {
+    described_class.notify_author_email(to_field: nil, subject_field: nil, decision_id: decision.id)
   }
 
   describe "#notify_author_email" do
-    it "sends email to the author's email" do
-      expect(email.to).to eq([paper.creator.email])
+    context 'with to field and subject field empty' do
+      it "sends email to the author's email" do
+        expect(email_with_no_fields_specified.to).to eq([paper.creator.email])
+      end
+
+      it "includes email subject" do
+        expect(email_with_no_fields_specified.subject).to eq "A decision has been registered on the manuscript, \"#{paper.title}\""
+      end
+
+      it "email body is paper.decision_letter, html escaped" do
+        expect(email_with_no_fields_specified.body).to include "Body text of a Decision Letter &lt;&lt; We Accept"
+      end
     end
 
-    it "includes email subject" do
-      expect(email.subject).to eq "A decision has been registered on the manuscript, \"#{paper.title}\""
-    end
+    context 'with to field and subject field populated with custom values' do
+      let(:email_to_arbitrary) {
+        described_class.notify_author_email(to_field: 'arb@example.com', subject_field: 'Your Submission', decision_id: decision.id)
+      }
+      it "sends email to the author's email" do
+        expect(email_to_arbitrary.to).to eq(['arb@example.com'])
+      end
 
-    it "email body is paper.decision_letter" do
-      expect(email.body.raw_source).to match(decision.letter)
+      it "sends email with a custom subject" do
+        expect(email_to_arbitrary.subject).to eq 'Your Submission'
+      end
     end
   end
 end