+++ spec/models/paper_spec.rb
@@ -417,7 +417,8 @@ describe Paper do
 
   context 'State Machine' do
     describe '#initial_submit' do
-      include_examples "transitions save state_updated_at", :initial_submit!
+      it_behaves_like "transitions save state_updated_at",
+        initial_submit: proc { paper.initial_submit! }
 
       it 'transitions to initially_submitted' do
         paper.initial_submit!
@@ -458,7 +459,8 @@ describe Paper do
     end
 
     describe '#submit!' do
-      include_examples "transitions save state_updated_at", :submit!
+      it_behaves_like "transitions save state_updated_at",
+        submit: proc { paper.submit!(paper.creator) }
 
       it 'does not transition when metadata tasks are incomplete' do
         expect(paper).to receive(:metadata_tasks_completed?).and_return(false)
@@ -541,32 +543,63 @@ describe Paper do
       end
     end
 
+    describe '#latest_withdrawal' do
+      let!(:joe) { FactoryGirl.create(:user) }
+      let!(:sally) { FactoryGirl.create(:user) }
+
+      before do
+        paper.withdraw!('reason 1', joe)
+        paper.reload.reactivate!
+        paper.withdraw!('reason 2', sally)
+      end
+
+      it 'returns the most recent withdrawal' do
+        withdrawal = paper.latest_withdrawal
+        expect(withdrawal).to be_kind_of(Withdrawal)
+        expect(withdrawal.withdrawn_by_user).to eq(sally)
+        expect(withdrawal.reason).to eq('reason 2')
+      end
+    end
+
     describe '#withdraw!' do
-      include_examples "transitions save state_updated_at", :withdraw!
+      let(:withdrawn_by_user) { FactoryGirl.build_stubbed(:user) }
+
+      it_behaves_like "transitions save state_updated_at",
+        withdraw: proc { paper.withdraw! 'A withdrawal reason', withdrawn_by_user }
 
       let(:paper) do
         FactoryGirl.create(:paper, :submitted, journal: journal)
       end
 
-      it "transitions to withdrawn without a reason" do
-        paper.withdraw!
+      it "raises an error when withdrawing without a reason and withdrawn_by_user" do
+        expect do
+          paper.withdraw!
+        end.to raise_error(ArgumentError, 'withdrawal_reason must be provided')
+
+        expect do
+          paper.withdraw! 'reason, but no user'
+        end.to raise_error(ArgumentError, 'withdrawn_by_user must be provided')
+      end
+
+      it 'withdraws the paper' do
+        paper.withdraw! 'reason', withdrawn_by_user
         expect(paper).to be_withdrawn
       end
 
       it "transitions to withdrawn with a reason" do
-        paper.withdraw! "Don't want to."
+        paper.withdraw! "Don't want to.", withdrawn_by_user
         expect(paper.withdrawn?).to eq true
       end
 
       it "marks the paper not editable" do
-        paper.withdraw!
+        paper.withdraw! "Some reason", withdrawn_by_user
         expect(paper).to_not be_editable
       end
     end
 
     describe '#invite_full_submission' do
-      include_examples "transitions save state_updated_at",
-                       :invite_full_submission!
+      it_behaves_like "transitions save state_updated_at",
+        invite_full_submission: proc { paper.invite_full_submission! }
 
       let(:paper) do
         FactoryGirl.create(:paper, :initially_submitted, journal: journal)
@@ -592,7 +625,8 @@ describe Paper do
     end
 
     describe '#reactivate!' do
-      include_examples "transitions save state_updated_at", :reactivate!
+      it_behaves_like "transitions save state_updated_at",
+        reactivate: proc { paper.reactivate! }
 
       let(:paper) do
         FactoryGirl.create(:paper, :submitted, journal: journal)
@@ -603,7 +637,7 @@ describe Paper do
       end
 
       before do
-        paper.withdraw!
+        paper.withdraw!('some reason', FactoryGirl.build_stubbed(:user))
       end
 
       it "transitions to the previous state" do
@@ -622,7 +656,7 @@ describe Paper do
       it "marks the paper with the previous editable state for unsubmitted papers" do
         paper = unsubmitted_paper
         expect(paper).to be_editable
-        paper.withdraw!
+        paper.withdraw!('some reason', FactoryGirl.build_stubbed(:user))
         expect(paper).to_not be_editable
         paper.reload.reactivate!
         expect(paper).to be_editable
@@ -631,7 +665,8 @@ describe Paper do
     end
 
     describe '#minor_check!' do
-      include_examples "transitions save state_updated_at", :minor_check!
+      it_behaves_like "transitions save state_updated_at",
+        minor_check: proc { paper.minor_check! }
 
       let(:paper) do
         FactoryGirl.create(:paper, :submitted, journal: journal)
@@ -652,8 +687,8 @@ describe Paper do
     end
 
     describe '#submit_minor_check!' do
-      include_examples "transitions save state_updated_at",
-                       :submit_minor_check!
+      it_behaves_like "transitions save state_updated_at",
+        submit_minor_check: proc { paper.submit_minor_check!(paper.creator) }
 
       let(:paper) do
         FactoryGirl.create(:paper, :submitted, journal: journal)
@@ -686,7 +721,8 @@ describe Paper do
           FactoryGirl.create(:paper, :submitted, journal: journal)
         end
 
-        include_examples "transitions save state_updated_at", :reject!
+        it_behaves_like "transitions save state_updated_at",
+          reject: proc { paper.reject! }
 
         it 'transitions to rejected state from submitted' do
           paper.reject!
@@ -703,7 +739,8 @@ describe Paper do
           )
         end
 
-        include_examples "transitions save state_updated_at", :reject!
+        it_behaves_like "transitions save state_updated_at",
+          reject: proc { paper.reject! }
 
         it 'transitions to rejected state from initially_submitted' do
           paper.reject!
@@ -713,7 +750,8 @@ describe Paper do
     end
 
     describe '#publish!' do
-      include_examples "transitions save state_updated_at", :publish!
+      it_behaves_like "transitions save state_updated_at",
+        publish: proc { paper.publish! }
 
       let(:paper) do
         FactoryGirl.create(:paper, :submitted, journal: journal)